Separated 99 Bottles of beer on the wall problem
